- Separated user count from municipality statistic - moved user count to top
